Function ovr_sys::audio::ovr_ReadWavFromBuffer

pub unsafe extern "C" fn ovr_ReadWavFromBuffer(
    outAudioChannel: *mut ovrAudioChannelData, 
    inputData: *const c_void, 
    dataSizeInBytes: c_int, 
    stereoChannelToUse: c_int
) -> ovrResult

Reads an audio channel from Wav (Waveform Audio File) data. Input must be a byte buffer representing a valid Wav file. Audio samples from the specified channel are read, converted to float [-1.0f, 1.0f] and returned through ovrAudioChannelData.

Supported formats: PCM 8b, 16b, 32b and IEEE float (little-endian only).

out outAudioChannel output audio channel data.

in inputData a binary buffer representing a valid Wav file data.

in dataSizeInBytes size of the buffer in bytes.

in stereoChannelToUse audio channel index to extract (0 for mono).