Module ovr_sys::directx

LibOVR functions for performing DirectX interop.

Functions

ovr_CreateMirrorTextureDX⚠
Create Mirror Texture which is auto-refreshed to mirror Rift contents produced by this application.

ovr_CreateTextureSwapChainDX⚠
Create Texture Swap Chain suitable for use with Direct3D 11 and 12.

ovr_GetMirrorTextureBufferDX⚠
Get the underlying buffer as any compatible COM interface (similar to QueryInterface)

ovr_GetTextureSwapChainBufferDX⚠
Get a specific buffer within the chain as any compatible COM interface (similar to QueryInterface)