Function ovr_sys::ovrPosef_FlipHandedness

pub unsafe extern "C" fn ovrPosef_FlipHandedness(
    inPose: *const ovrPosef, 
    outPose: *mut ovrPosef
)

Tracking poses provided by the SDK come in a right-handed coordinate system. If an application is passing in ovrProjection_LeftHanded into ovrMatrix4f_Projection, then it should also use this function to flip the HMD tracking poses to be left-handed.

While this utility function is intended to convert a left-handed ovrPosef into a right-handed coordinate system, it will also work for converting right-handed to left-handed since the flip operation is the same for both cases.

in inPose that is right-handed

out outPose that is requested to be left-handed (can be the same pointer to inPose)