Function ovr_sys::ovr_CommitTextureSwapChain

pub unsafe extern "C" fn ovr_CommitTextureSwapChain(
    session: ovrSession, 
    chain: ovrTextureSwapChain
) -> ovrResult

Commits any pending changes to an ovrTextureSwapChain, and advances its current index

in session Specifies an ovrSession previously returned by ovr_Create.

in chain Specifies the ovrTextureSwapChain to commit.

Note: When Commit is called, the texture at the current index is considered ready for use by the runtime, and further writes to it should be avoided. The swap chain's current index is advanced, providing there's room in the chain. The next time the SDK dereferences this texture swap chain, it will synchronize with the app's graphics context and pick up the submitted index, opening up room in the swap chain for further commits.

Returns an ovrResult for which OVR_SUCCESS(result) is false upon error. Failures include but aren't limited to:

ovrError_TextureSwapChainFull: ovr_CommitTextureSwapChain was called too many times on a texture swapchain without calling submit to use the chain.

see ovr_CreateTextureSwapChainDX, ovr_CreateTextureSwapChainGL, ovr_CreateTextureSwapChainVk