Function ovr_sys::ovr_GetEyePoses

pub unsafe extern "C" fn ovr_GetEyePoses(
    session: ovrSession, 
    frameIndex: c_longlong, 
    latencyMarker: ovrBool, 
    hmdToEyeOffset: *const [ovrVector3f; 2], 
    outEyePoses: *const [ovrPosef; 2], 
    outSensorSampleTime: *mut f64
)

Returns the predicted head pose in outHmdTrackingState and offset eye poses in outEyePoses.

This is a thread-safe function where caller should increment frameIndex with every frame and pass that index where applicable to functions called on the rendering thread. Assuming outEyePoses are used for rendering, it should be passed as a part of ovrLayerEyeFov. The caller does not need to worry about applying HmdToEyeOffset to the returned outEyePoses variables.

in hmd Specifies an ovrSession previously returned by ovr_Create.

in frameIndex Specifies the targeted frame index, or 0 to refer to one frame after the last time ovr_SubmitFrame was called.

in latencyMarker Specifies that this call is the point in time where the "App-to-Mid-Photon" latency timer starts from. If a given ovrLayer provides "SensorSampleTimestamp", that will override the value stored here.

in hmdToEyeOffset Can be ovrEyeRenderDesc.HmdToEyeOffset returned from ovr_GetRenderDesc. For monoscopic rendering, use a vector that is the average of the two vectors for both eyes.

out outEyePoses The predicted eye poses.

out outSensorSampleTime The time when this function was called. May be NULL, in which case it is ignored.