Function ovr_sys::ovr_GetTrackerDesc

pub unsafe extern "C" fn ovr_GetTrackerDesc(
    session: ovrSession, 
    trackerDescIndex: c_uint
) -> ovrTrackerDesc

Returns a given attached tracker description.

ovr_Initialize must have first been called in order for this to succeed, otherwise the returned trackerDescArray will be zero-initialized. The data returned by this function can change at runtime.

session Specifies an ovrSession previously returned by ovr_Create.

trackerDescIndex Specifies a tracker index. The valid indexes are in the range of 0 to the tracker count returned by ovr_GetTrackerCount.

Returns ovrTrackerDesc. An empty ovrTrackerDesc will be returned if trackerDescIndex is out of range.

see ovrTrackerDesc, ovr_GetTrackerCount