Function ovr_sys::ovr_GetTrackingState

pub unsafe extern "C" fn ovr_GetTrackingState(
    session: ovrSession, 
    absTime: f64, 
    latencyMarker: ovrBool
) -> ovrTrackingState

Returns tracking state reading based on the specified absolute system time.

Pass an absTime value of 0.0 to request the most recent sensor reading. In this case both PredictedPose and SamplePose will have the same value.

This may also be used for more refined timing of front buffer rendering logic, and so on.

This may be called by multiple threads.

session Specifies an ovrSession previously returned by ovr_Create.

absTime Specifies the absolute future time to predict the return ovrTrackingState value. Use 0 to request the most recent tracking state.

latencyMarker Specifies that this call is the point in time where the "App-to-Mid-Photon" latency timer starts from. If a given ovrLayer provides "SensorSampleTime", that will override the value stored here.

Returns the ovrTrackingState that is predicted for the given absTime.

see ovrTrackingState, ovr_GetEyePoses, ovr_GetTimeInSeconds