Function ovr_sys::ovr_IdentifyClient

pub unsafe extern "C" fn ovr_IdentifyClient(
    identity: *const c_char
) -> ovrResult

Identify client application info.

The string is one or more newline-delimited lines of optional info indicating engine name, engine version, engine plugin name, engine plugin version, engine editor. The order of the lines is not relevant. Individual lines are optional. A newline is not necessary at the end of the last line.

Call after ovr_Initialize and before the first call to ovr_Create.

Each value is limited to 20 characters. Key names such as 'EngineName:', 'EngineVersion:' do not count towards this limit.

identity Specifies one or more newline-delimited lines of optional info:

EngineName: %s\n
EngineVersion: %s\n
EnginePluginName: %s\n
EnginePluginVersion: %s\n
EngineEditor: <boolean> ('true' or 'false')\n

Example code

let identity = b"EngineName: Unity\n\
                 EngineVersion: 5.3.3\n\
                 EnginePluginName: `OVRPlugin`\n\
                 EnginePluginVersion: 1.2.0\n\
                 EngineEditor: true\0";
ovr_IdentifyClient(CStr::from_bytes_with_nul_unchecked(identity).as_ptr());