Function ovr_sys::ovr_TestBoundary

pub unsafe extern "C" fn ovr_TestBoundary(
    session: ovrSession, 
    deviceBitmask: ovrTrackedDeviceType, 
    boundaryType: ovrBoundaryType, 
    outTestResult: *mut ovrBoundaryTestResult
) -> ovrResult

Tests collision/proximity of position tracked devices (e.g. HMD and/or Touch) against the Boundary System. Note: this method is similar to ovr_BoundaryTestPoint but can be more precise as it may take into account device acceleration/momentum.

session Specifies an ovrSession previously returned by ovr_Create.

deviceBitmask Bitmask of one or more tracked devices to test.

boundaryType Must be either ovrBoundary_Outer or ovrBoundary_PlayArea.

out outTestResult Result of collision/proximity test, contains information such as distance and closest point. Returns an ovrResult for which OVR_SUCCESS(result) is false upon error and true upon success. Return values include but aren't limited to: * ovrSuccess: The call succeeded and a result was returned. * ovrSuccess_BoundaryInvalid: The call succeeded but the result is not a valid boundary due to not being set up. * ovrSuccess_DeviceUnavailable: The call succeeded but the device referred to by deviceBitmask is not available.

see ovrBoundaryTestResult