Module ovr_sys::opengl

LibOVR functions for performing OpenGL interop.

Functions

ovr_CreateMirrorTextureGL⚠
Creates a Mirror Texture which is auto-refreshed to mirror Rift contents produced by this application.

ovr_CreateTextureSwapChainGL⚠
Creates a TextureSwapChain suitable for use with OpenGL.

ovr_GetMirrorTextureBufferGL⚠
Get a the underlying buffer as a GL texture name

ovr_GetTextureSwapChainBufferGL⚠
Get a specific buffer within the chain as a GL texture name