Struct ovr_sys::ovrPerfStatsPerCompositorFrame

#[repr(C)]

pub struct ovrPerfStatsPerCompositorFrame {
    pub _align: [u32; 0],
    pub HmdVsyncIndex: c_int,
    pub AppFrameIndex: c_int,
    pub AppDroppedFrameCount: c_int,
    pub AppMotionToPhotonLatency: f32,
    pub AppQueueAheadTime: f32,
    pub AppCpuElapsedTime: f32,
    pub AppGpuElapsedTime: f32,
    pub CompositorFrameIndex: c_int,
    pub CompositorDroppedFrameCount: c_int,
    pub CompositorLatency: f32,
    pub CompositorCpuElapsedTime: f32,
    pub CompositorGpuElapsedTime: f32,
    pub CompositorCpuStartToGpuEndElapsedTime: f32,
    pub CompositorGpuEndToVsyncElapsedTime: f32,
}

Contains the performance stats for a given SDK compositor frame

All of the c_int fields can be reset via the ovr_ResetPerfStats call.

Fields
_align: [u32; 0] HmdVsyncIndex: c_int

Vsync Frame Index - increments with each HMD vertical synchronization signal (i.e. vsync or refresh rate) If the compositor drops a frame, expect this value to increment more than 1 at a time.

AppFrameIndex: c_int

Application stats

Index that increments with each successive ovr_SubmitFrame call

AppDroppedFrameCount: c_int

If the app fails to call ovr_SubmitFrame on time, then expect this value to increment with each missed frame

AppMotionToPhotonLatency: f32

Motion-to-photon latency for the application This value is calculated by either using the SensorSampleTime provided for the ovrLayerEyeFov or if that is not available, then the call to ovr_GetTrackingState which has latencyMarker set to ovrTrue

AppQueueAheadTime: f32

Amount of queue-ahead in seconds provided to the app based on performance and overlap of CPU & GPU utilization A value of 0.0 would mean the CPU & GPU workload is being completed in 1 frame's worth of time, while 11 ms (on the CV1) of queue ahead would indicate that the app's CPU workload for the next frame is overlapping the app's GPU workload for the current frame.

AppCpuElapsedTime: f32

Amount of time in seconds spent on the CPU by the app's render-thread that calls ovr_SubmitFrame Measured as elapsed time between from when app regains control from ovr_SubmitFrame to the next time the app calls ovr_SubmitFrame.

AppGpuElapsedTime: f32

Amount of time in seconds spent on the GPU by the app Measured as elapsed time between each ovr_SubmitFrame call using GPU timing queries.

CompositorFrameIndex: c_int

SDK Compositor stats

Index that increments each time the SDK compositor completes a distortion and timewarp pass Since the compositor operates asynchronously, even if the app calls ovr_SubmitFrame too late, the compositor will kick off for each vsync.

CompositorDroppedFrameCount: c_int

Increments each time the SDK compositor fails to complete in time This is not tied to the app's performance, but failure to complete can be tied to other factors such as OS capabilities, overall available hardware cycles to execute the compositor in time and other factors outside of the app's control.

CompositorLatency: f32

Motion-to-photon latency of the SDK compositor in seconds This is the latency of timewarp which corrects the higher app latency as well as dropped app frames.

CompositorCpuElapsedTime: f32

The amount of time in seconds spent on the CPU by the SDK compositor. Unless the VR app is utilizing all of the CPU cores at their peak performance, there is a good chance the compositor CPU times will not affect the app's CPU performance in a major way.

CompositorGpuElapsedTime: f32

The amount of time in seconds spent on the GPU by the SDK compositor. Any time spent on the compositor will eat away from the available GPU time for the app.

CompositorCpuStartToGpuEndElapsedTime: f32

The amount of time in seconds spent from the point the CPU kicks off the compositor to the point in time the compositor completes the distortion & timewarp on the GPU. In the event the GPU time is not available, expect this value to be -1.0f

CompositorGpuEndToVsyncElapsedTime: f32

The amount of time in seconds left after the compositor is done on the GPU to the associated V-Sync time.

In the event the GPU time is not available, expect this value to be -1.0f