Module ovr_sys::vulkan

LibOVR functions for performing Vulkan interop.

Functions

ovr_CreateMirrorTextureWithOptionsVk⚠
Create Mirror Texture which is auto-refreshed to mirror Rift contents produced by this application.

ovr_CreateTextureSwapChainVk⚠
Create Texture Swap Chain suitable for use with Vulkan

ovr_GetMirrorTextureBufferVk⚠
Get a the underlying mirror Image

ovr_GetSessionPhysicalDeviceVk⚠
Find PhysicalDevice matching ovrGraphicsLuid

ovr_GetTextureSwapChainBufferVk⚠
Get a specific Image within the chain

ovr_SetSynchonizationQueueVk⚠
Select Queue to block on till rendering is complete