Function ovr_sys::vulkan::ovr_GetSessionPhysicalDeviceVk

pub unsafe extern "C" fn ovr_GetSessionPhysicalDeviceVk(
    session: ovrSession, 
    luid: ovrGraphicsLuid, 
    instance: Instance, 
    out_physicalDevice: *mut PhysicalDevice
) -> ovrResult

Find PhysicalDevice matching ovrGraphicsLuid

in session Specifies an ovrSession previously returned by ovr_Create.

in luid Specifies the luid returned from ovr_Create.

in instance Specifies an Instance to search for matching luids in.

out out_physicalDevice Returns the PhysicalDevice matching the instance and luid.

Returns an ovrResult indicating success or failure. In the case of failure, use ovr_GetLastErrorInfo to get more information.

Note: This function enumerates the current physical devices and returns the one matching the luid. It must be called at least once prior to any ovr_CreateTextureSwapChainVk or ovr_CreateMirrorTextureWithOptionsVk calls, and the instance must remain valid for the lifetime of the returned objects. It is assumed the Device created by the application will be for the returned physical device.