Function ovr_sys::vulkan::ovr_SetSynchonizationQueueVk

pub unsafe extern "C" fn ovr_SetSynchonizationQueueVk(
    session: ovrSession, 
    queue: Queue
) -> ovrResult

Select Queue to block on till rendering is complete

in session Specifies an ovrSession previously returned by ovr_Create.

in queue Specifies a Queue to add a Fence operation to and wait on.

Returns an ovrResult indicating success or failure. In the case of failure, use ovr_GetLastErrorInfo to get more information.

Note: The queue may be changed at any time but only the value at the time ovr_SubmitFrame is called will be used. ovr_SetSynchonizationQueueVk must be called with a valid Queue created on the same Device the texture sets were created on prior to the first call to ovr_SubmitFrame. An internally created Fence object will be signalled by the completion of operations on queue and waited on to synchronize the VR compositor.